Rebuild the Biosynthetic Path of a Glycan

Reconstruct the biosynthetic pathway of a glycan from a given starting structure using enzymatic reactions. This function uses a combination of depth-first search and heuristics to find a likely pathway, or all feasible pathways.

Usage

rebuild_biosynthesis(
  glycan,
  enzymes = NULL,
  max_steps = 20,
  filter = NULL,
  return = c("shortest", "all")
)

Arguments

glycan

A glyrepr::glycan_structure() scalar, or a character string supported by glyparse::auto_parse().

enzymes

A character vector of gene symbols, or a list of enzyme() objects. If NULL (default), all available enzymes will be used.

max_steps

Integer, maximum number of enzymatic steps to search. Default is 20.

filter

Optional function to filter generated glycans at each step. Should take a glyrepr::glycan_structure() vector as input and return a logical vector of the same length.

return

One of "shortest" or "all". If "shortest", returns the first shortest path found. If "all", returns a graph containing all possible paths within max_steps.

Value

An igraph::igraph() object representing the synthesis path(s). Vertices represent glycan structures with name attribute containing IUPAC-condensed strings. Edges represent enzymatic reactions with enzyme attribute containing gene symbols and step attribute indicating the step number.

Details

For N-glycans, the starting structure is assumed to be "Glc(3)Man(9)GlcNAc(2)", the N-glycan precursor transfered to Asn by OST. For O-glycans, the starting structure is assumed to be "GalNAc(a1-".

Examples

library(glyrepr)
library(glyparse)

# Rebuild the biosynthetic pathway of a glycan
glycan <- "Neu5Ac(a2-3)Gal(b1-4)[Fuc(a1-3)]GlcNAc(b1-3)Gal(b1-3)GalNAc(a1-"
path <- rebuild_biosynthesis(glycan, max_steps = 20)

# View the path
igraph::as_data_frame(path, what = "edges")
#>                                                  from
#> 1                                          GalNAc(a1-
#> 2                                 Gal(b1-3)GalNAc(a1-
#> 3                     GlcNAc(b1-3)Gal(b1-3)GalNAc(a1-
#> 4            Gal(b1-4)GlcNAc(b1-3)Gal(b1-3)GalNAc(a1-
#> 5 Fuc(a1-3)[Gal(b1-4)]GlcNAc(b1-3)Gal(b1-3)GalNAc(a1-
#>                                                                to  enzyme step
#> 1                                             Gal(b1-3)GalNAc(a1- C1GALT1    1
#> 2                                 GlcNAc(b1-3)Gal(b1-3)GalNAc(a1-  B3GNT3    2
#> 3                        Gal(b1-4)GlcNAc(b1-3)Gal(b1-3)GalNAc(a1- B4GALT1    3
#> 4             Fuc(a1-3)[Gal(b1-4)]GlcNAc(b1-3)Gal(b1-3)GalNAc(a1-    FUT3    4
#> 5 Neu5Ac(a2-3)Gal(b1-4)[Fuc(a1-3)]GlcNAc(b1-3)Gal(b1-3)GalNAc(a1- ST3GAL3    5