Enzymes

Glycosylation is catalyzed by two types of enzymes: glycosyltransferases (GTs) and glycoside hydrolases (GHs). Glycosyltransferases catalyze the transfer of a sugar residue from a donor to an acceptor. Glycoside hydrolases catalyze the removal of a sugar residue from a substrate. glyenzy provides a data structure (glyenzy_enzyme) to represent these enzymes.

Use enzyme() with a gene symbol to load a predefined enzyme. For example, use enzyme("ST3GAL3") to load the enzyme ST3GAL3.

Throughout the package, you can use enzyme()s for any enzyme argument, or just use the gene symbol directly. For example, involve("Neu5Ac(a2-3)Gal(b1-3)GalNAc(a1-", "ST3GAL3") and involve("Neu5Ac(a2-3)Gal(b1-3)GalNAc(a1-", enzyme("ST3GAL3")) are equivalent.

Usage

enzyme(symbol)

Arguments

symbol

The gene symbol of the enzyme.

Value

A glyenzy_enzyme object.

Explanation about glyenzy_enzyme

An enzyme() is a list with the following elements:

name: the name of the enzyme, usually the gene symbol.

rules: a list of glyenzy_enzyme_rule objects. Each rule is a list with the following fields:

acceptor: the motif that the enzyme recognizes

acceptor_alignment: the alignment of the acceptor

rejects: the motifs that the enzyme rejects to act on

product: the product generated by the enzyme

acceptor_idx: the node index of the acceptor where the enzyme acts on. For GTs, this is the node new residue is attached to. For GHs, this is the node that is removed.

product_idx: the node index of the product residue in the product structure. For GTs, this is the index of the newly added residue in the product. For GHs, this is NULL (no new residue is added).

new_residue: the new residue added by the enzyme. For GHs, this is NULL.

new_linkage: the linkage of the new residue. For GHs, this is NULL.

type: the type of the enzyme, "GT" for glycosyltransferase or "GH" for glycoside hydrolase.

species: the species of the enzyme, e.g. "human" or "mouse".

You can see all these information by printing the enzyme object.

Examples

library(glyrepr)

enzyme("ST3GAL3")
#> 
#> ── Enzyme: ST3GAL3 ─────────────────────────────────────────────────────────────
#> ℹ Type: "GT" (Glycosyltransferase)
#> ℹ Species: "human"
#> 
#> ── Rules (2) ──
#> 
#> → Rule 1: terminal alignment
#> Acceptor: "Gal(b1-3)GlcNAc(b1-"
#> Product: "Neu5Ac(a2-3)Gal(b1-3)GlcNAc(b1-"
#> → Rule 2: terminal alignment
#> Acceptor: "Gal(b1-4)GlcNAc(b1-"
#> Product: "Neu5Ac(a2-3)Gal(b1-4)GlcNAc(b1-"